+++ app/src/main/java/com/ddd/attendance/feature/admin/screen/ScheduleScreen.kt
@@ -1,0 +1,4 @@
+package com.ddd.attendance.feature.admin.screen
+
+class ScheduleScreen {
+}
